requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.24
jinja2>=3.1
pydantic>=2.5
pydantic-settings>=2.2
SQLAlchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.12
python-multipart>=0.0.9
passlib[bcrypt]>=1.7
bcrypt>=4.0
python-jose>=3.3
PyYAML>=6.0
loguru>=0.7
tenacity>=8.2
